+++ js/get-check-length-str.js
@@ -1,0 +1,3 @@
+const getCheckLengthStr = (str, lengthStr) => str.length <= lengthStr;
+
+export {getCheckLengthStr};
