+++ js/get-random-string.js
@@ -1,0 +1,5 @@
+import {getRandomInt} from './util.js';
+
+const getRandomString = (str) => str.substr(0, getRandomInt(1, 10));
+
+export {getRandomString};
