+++ js/util.js
@@ -1,0 +1,16 @@
+function getRandomInt(min, max) {
+  min = Math.ceil(min);
+  max = Math.floor(max);
+
+  if (min < 0 || max < 0 || typeof min !== 'number' || typeof max !== 'number') {
+    return NaN;
+  } else if (min > max) {
+    const newMax = min;
+    min = max;
+    max = newMax;
+  }
+
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+}
+
+export {getRandomInt};
